pyproject.toml
[project]
name = "humiPy"
dynamic = ["version"]
requires-python = ">= 3.11"
dependencies = [
    "requests==2.31.0",
    "python-dotenv==1.0.0",
    "python-dateutil==2.8.2",
]